Fixed Integration Method. Debugging in GyroY:Merker ind GyroX DX

+++ src/ringbuffer.h
@@ -75,7 +75,7 @@ int32_t moving_average (common_buffer_data* buffer) {
 int32_t integration(common_buffer_data* buffer,uint32_t * speed, int32_t accel_linear,int32_t accel_linear_last_value) {
     uint32_t dt = buffer->current_time-buffer->last_time;
     int32_t dx = accel_linear-accel_linear_last_value;
-    *speed = *speed+((dx*dt)/2)+(accel_linear_last_value*dt);
+    *speed = *speed+((dx*(int32_t)dt)/2)+(accel_linear_last_value*dt);
     return dx;
 }
 double scaling (int32_t* buffer_sum)
